Fix for the issue #1480 : make documentation targets near to SALOME common rules

+++ src/Config/Config_DataModelReader.h
@@ -36,7 +36,7 @@ class Config_DataModelReader : public Config_XMLReader
   CONFIG_EXPORT std::string rootType() const { return myRootTypes; }
 
   /// Returns number of folders under root 
-  CONFIG_EXPORT int rootFoldersNumber() const { return int(myRootFolderNames.size()); }
+  CONFIG_EXPORT size_t rootFoldersNumber() const { return myRootFolderNames.size(); }
 
   /// Returns name of the folder by its Id
   /// \param theId id of the folder
@@ -68,7 +68,7 @@ class Config_DataModelReader : public Config_XMLReader
   CONFIG_EXPORT std::string subType() const { return mySubTypes; }
 
   /// Returns number of folders under sub document 
-  CONFIG_EXPORT int subFoldersNumber() const { return int(mySubFolderNames.size()); }
+  CONFIG_EXPORT size_t subFoldersNumber() const { return mySubFolderNames.size(); }
 
   /// Returns name of the folder by its Id
   /// \param theId id of the folder
